backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "capacity test"
authors = ["Ismael"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
